# Optimize IP comparison