#38 has some ideas to speed up library. 
Implemented ideas: 
- LINQ extensions => for loop
- using offset for  IPv4 mapped to IPv6

Gained performance 
- Bit.GE : 82.1% time less - 3M calls -  24534 ms -> 4400 ms
- Bit.LE :  81.9% time less - 4 999 911 calls - 33004 ms -> 7063 ms

**Tests** 
IPv4 range parse
- 1M calls
- gained 5.42%

IPv6 range parse:
- 999 908 calls
- gained 13.85%

IPv6 Contains - called both LE and GE
- 1M calls
- gained 9.10%

IPv4 Contains - called both LE and GE
- 1M calls
- gained 2.91%

IPv4 mapped to IPv6 Contains - called both LE and GE
- 1M calls
- gained 10.22%

Not implemented: 
3. I'm not fully sure it's a good idea.
4.  Get hash code takes for IPv4 as the most significant byte the last octet. `1.0.0.255` > `1.0.1.0`