@jsakamoto 2018-04-23
Thank you for your contribution!

I'll merge this pull request due to I completed reviewing this pull request.

By the way, this pull request include the changes below:

- Build binary is not compatible with lower version. (ex. `Bits.GE(arg0, arg1)` is missing.)
- This pull request will fix issue #39 (thank you!), but it will causes broken exist programs those depend on this bug.

Therefore, I'll merge this pull request manually with a few steps for keeping backward compatible.  
So, for this work, please give me more few days.